Lab Part 2 - Added error checking for delete CATEGORY such that a category will NOT be deleted if products in that category are in DB.

+++ delete_category.php
@@ -8,14 +8,29 @@ if ($category_id == null || $category_id == false) {
 	include('error.php');
 } else {
 	require_once('database.php');
-// Delete the product from the database
-    $query = 'DELETE FROM categories
-              WHERE categoryID = :category_id';
-    $statement = $db->prepare($query);
-    $statement->bindValue(':category_id', $category_id);
-    $success = $statement->execute();
-    $statement->closeCursor();    
+	
+	//See how many rows are in products table with the given categoryID
+	$query = 'SELECT * FROM products
+			 WHERE categoryID = :category_id';
+	$statement = $db->prepare($query);
+	$statement->bindValue(':category_id', $category_id);
+	$statement->execute();
+	$products = $statement->fetchAll();
+	$statement->closeCursor();
+	if (count($products) > 0) {
+		$error = "The category cannot be deleted since it contains products.";
+		include('error.php');
+	} else {
+	// Delete the category from the database
+		$query = 'DELETE FROM categories
+	              WHERE categoryID = :category_id';
+	    $statement = $db->prepare($query);
+	    $statement->bindValue(':category_id', $category_id);
+	    $statement->execute();
+	    $statement->closeCursor(); 
+	    
+	    // Display the Product List page
+	    include('category_list.php');
+    }
 }
 
-// Display the Product List page
-include('category_list.php');
